// Nível 09 - Node.js com Containers/08 - Desafio Prático - Gerenciador de Tarefas/task-manager/src/schemas/auth-schemas.ts
import { z } from "zod";

export const registerSchema = z.object({
  name: z.string().min(2),
  email: z.string().email(),
  password: z.string().min(6),
  role: z.enum(["admin", "member"]).optional(),
});

export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string().min(6),
});